# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Adapta API
  description: |
    **Your model. Your data. Your servers.**

    Adapta is a self-hosted model-customization platform. Technical teams deploy
    it on their own infrastructure to specialize and serve private language models two ways,
    each behind an OpenAI-compatible API. Your data never leaves your servers.

    **Give it knowledge (RAG)** — upload documents, embed them, and get answers grounded in
    your content, with citations. Weights never change. CPU-only.

    **Change how it behaves (fine-tuning / LoRA)** — train a QLoRA adapter on instruction
    data; the adapter only serves once it passes the eval gate (score ≥ 0.6 or a clear
    improvement over the base model). Requires a GPU.

    **Composed endpoints** — a single project endpoint can serve both: retrieved context
    with citations (RAG) and trained tone/format (the LoRA adapter) in one call.

    ---

    **Auth:** management endpoints require a JWT Bearer token from `POST /v1/auth/login`.
    The `POST /v1/chat/completions` endpoint uses a scoped project API key (`adp_*` prefix).

  version: 0.1.0
  license:
    name: MIT
    url: https://github.com/yielab/adapta/blob/main/LICENSE
  contact:
    name: Adapta
    url: https://github.com/yielab/adapta
  x-logo:
    url: /console/logo.svg
    altText: Adapta
servers:
  - url: /v1
    description: Versioned API
...
paths:
...
  /projects/{project_id}/files:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a document for indexing (both project types)
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: {type: string, format: binary}
              required: [file]
      responses:
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '202': {description: Accepted — indexing in background}

    get:
      operationId: listFiles
      tags: [files]
      summary: List uploaded files for a project
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Files, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/FileResponse'}}}}}
  /projects/{project_id}/files/{file_id}:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file and remove its chunks from the collection
      security: [{BearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
        - {name: file_id, in: path, required: true, schema: {type: string}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '204': {description: Deleted}

  # ──────────────────────────────────────────────────────────────────────────
  # Datasets (fine-tune)
  # ──────────────────────────────────────────────────────────────────────────
...
  /projects/{project_id}/jobs:
    post:
      operationId: createJob
      tags: [jobs]
      summary: Enqueue a training job
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/JobCreateRequest'}
      responses:
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '202': {description: Queued, content: {application/json: {schema: {$ref: '#/components/schemas/JobResponse'}}}}

    get:
      operationId: listJobs
      tags: [jobs]
      summary: List training jobs for a project
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Jobs, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/JobResponse'}}}}}
  /projects/{project_id}/jobs/{job_id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job status and live progress
      security: [{BearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
        - {name: job_id, in: path, required: true, schema: {type: string}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Job, content: {application/json: {schema: {$ref: '#/components/schemas/JobResponse'}}}}

  # ──────────────────────────────────────────────────────────────────────────
  # Endpoint
  # ──────────────────────────────────────────────────────────────────────────
...
  /projects/{project_id}/keys:
    post:
      operationId: createKey
      tags: [keys]
      summary: Create a scoped API key for the project endpoint
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/KeyCreateRequest'}
      responses:
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '201': {description: Key (shown once), content: {application/json: {schema: {$ref: '#/components/schemas/KeyCreatedResponse'}}}}

    get:
      operationId: listKeys
      tags: [keys]
      summary: List API keys (prefix only — full key never returned after creation)
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Keys, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/KeyResponse'}}}}}
  /projects/{project_id}/keys/{key_id}:
    delete:
      operationId: revokeKey
      tags: [keys]
      summary: Revoke a key
      security: [{BearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
        - {name: key_id, in: path, required: true, schema: {type: string}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '204': {description: Revoked}

  # ──────────────────────────────────────────────────────────────────────────
  # Dataset synthesis (finetune projects only)
  # ──────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: Scoped project API key (adp_...)
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema: {type: string}
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
...
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
...
    FileResponse:
      type: object
      required: [id, filename, status, num_chunks, size_bytes, uploaded_at]
      properties:
        id: {type: string}
        filename: {type: string}
        status: {type: string}
        num_chunks: {type: integer, nullable: true}
        size_bytes: {type: integer}
        uploaded_at: {type: string}

    # Datasets
...
    TrainingConfigInput:
      # Bounds mirror the server's enqueue-time validation (A4.12,
      # adapta/services/training.py _HYPERPARAM_BOUNDS) — out-of-range
      # values are rejected with 422 before a job is created.
      type: object
      properties:
        num_epochs: {type: integer, default: 3, minimum: 1, maximum: 100}
        batch_size: {type: integer, default: 4, minimum: 1, maximum: 128}
        learning_rate: {type: number, default: 0.0002, minimum: 1.0e-6, maximum: 0.1}
        lora_r: {type: integer, default: 16, minimum: 1, maximum: 256}
        lora_alpha: {type: integer, default: 32, minimum: 1, maximum: 512}
        lora_dropout: {type: number, default: 0.1, minimum: 0.0, maximum: 0.9}
        max_seq_length: {type: integer, default: 512, minimum: 16, maximum: 8192}
    JobCreateRequest:
      type: object
      required: [dataset_id]
      properties:
        dataset_id: {type: string}
        training_config: {$ref: '#/components/schemas/TrainingConfigInput', nullable: true}
    JobResponse:
      type: object
      required: [id, status, progress, logs, adapter_path, eval_score, eval_passed, eval_metrics, error_message, created_at]
      properties:
        id: {type: string}
        status: {$ref: '#/components/schemas/JobStatus'}
        progress: {type: number, minimum: 0, maximum: 1}
        logs: {type: string, nullable: true}
        adapter_path: {type: string, nullable: true}
        eval_score: {type: number, nullable: true}
        eval_passed: {type: boolean, nullable: true}
        # Full eval result for an auditable gate verdict (A4.6): score, base_score,
        # score_delta, held_out, metrics, sample_predictions. Null until evaluation.
        eval_metrics: {type: object, nullable: true, additionalProperties: true}
        error_message: {type: string, nullable: true}
        created_at: {type: string}

    # Endpoint
...
    KeyCreateRequest:
      type: object
      required: [name]
      properties:
        # maxLength mirrors api_keys.name String(128) — see RegisterRequest note.
        name: {type: string, minLength: 1, maxLength: 128}
    KeyCreatedResponse:
      type: object
      required: [id, name, key, prefix]
      properties:
        id: {type: string}
        name: {type: string}
        key: {type: string, description: "Full key — shown once only"}
        prefix: {type: string}
    KeyResponse:
      type: object
      required: [id, name, prefix, is_active, created_at]
      properties:
        id: {type: string}
        name: {type: string}
        prefix: {type: string}
        is_active: {type: boolean}
        created_at: {type: string}

    # Synthesis
...
    Error:
      type: object
      properties:
        code: {type: string}
        message: {type: string}
        correlation_id: {type: string, nullable: true}
    ErrorResponse:
      type: object
      properties:
        error: {$ref: '#/components/schemas/Error'}